Refactor TrainingHistory into module
